+++ VVCGithubUsers/UsersList/UsersTableViewModel.swift
@@ -13,7 +13,7 @@ import CoreData
 class UsersTableViewModel {
   var cellViewModels: [CellItemable] = []
 
-  private var lastUserID: Int16 = 0
+  private var lastUserID: Int64 = 0
   private(set) var selectedUser: CellItemable?
 
   private(set) var profileViewModel: ProfileViewModel?
@@ -105,7 +105,7 @@ class UsersTableViewModel {
 
 protocol CellItemable {
   var userName: String { get set }
-  var id: Int16 { get set }
+  var id: Int64 { get set }
   var details: String { get set }
 
   func cellInstance(tableView: UITableView, indexPath: IndexPath) -> UITableViewCell
